added update optimisations and image resizing for gallery.
Gallery widget now finds the greater value of an images height and
width, and resizes it to match the dimensions of the frame whilst
keeping the smaller value proportional.

Changes were made to how the widgets update in tkinter, having each
function on it's own update cycle instead of all looping in a single
update function.

Part of this is increasing the update rate on the system and processes,
but slowing down the update rate on the stars and notes to hopefully
save resources on database reads.