adding javascript forms and database connections for stars.
the stars page, until now has no way of setting which user the "star"
would be set to, and the old javascript that made the majority of forms
the site used didn't have a dropdown list input. The stars page now
posts to itself with a specific database connection that is properly
prepared and (hopefully) properly escaped.