Beginning of new forms and internal POST processing.
Stars no posts to itself and uses a new form generated in javascript.
Javascript now creates forms in a different way, generating a set form
per page instead of just doing whatever javascript function throws at
it.
updated stars on the sidebar to show a per-user basis of stars.
schema changed to create rewards with a "link" column to store an url of
a shop or something.