Dash style change and rules creation.
fix to image_upload.php because it needed username but username wasn't
in POST data.

Change in ironserver.php to fix a bad error handler.

rules now has a "new" form that allows you to choose who a rule applies
to ( NEARLY FINISHED!!)

Got rid of the star on rewards, it messed with javascript getting info.

SQL schema changed to set 'applies_to' as TEXT not INT