Change rules to use user id instead of usernames
Using user id in the applies to column lets me fetch the username from
the users table, this means that when running "edit username" I don't
have to run through the applies to column and take the list out, make it
an array, edit the username, make it a string and put it back in... for
every single row.

I also changed settings to use more prepared statements.